requirements.txt
APScheduler==2.1.2
backports.functools-lru-cache==1.5
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cheroot==6.5.6
CherryPy==8.5.0
DBUtils==1.3
decorator==4.4.0
gevent==1.4.0
gnsq==1.0.1
greenlet==0.4.15
idna==2.8
jaraco.functools==2.0
more-itertools==7.2.0
NIM-API-Python==0.1.0
nsq-py==0.1.11
Pillow==6.1.0
portend==2.5
publicsuffix==1.1.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pysnowflake==0.1.3
pytz==2019.2
qiniu==7.2.6
qrcode==6.1
rc4==0.2
redis==3.3.8
redis-py-cluster==2.0.0
redisco==0.1.4
requests==2.22.0
six==1.12.0
statsd==3.3.0
tempora==1.14.1
tornado==6.0.3
tzlocal==2.0.0
url==0.4.2
urllib3==1.25.3
zc.lockfile==2.0
